e2e reciliency

+++ test/e2e/environment/kind-operator/dualstack/test.sh
@@ -18,12 +18,15 @@ function new_vip () {
 
 function new_vip_revert () {
     kubectl apply -f $(dirname -- $(readlink -fn -- "$0"))/configuration/init-trench-a.yaml
+    if [ ! $? -eq 0 ]; then exit 1 ; fi
     kubectl delete vip -n red vip-a-2-v4
+    if [ ! $? -eq 0 ]; then exit 1 ; fi
     kubectl delete vip -n red vip-a-2-v6
 }
 
 function delete_create_trench () {
     kubectl delete trench trench-a -n red
+    if [ ! $? -eq 0 ]; then exit 1 ; fi
     sleep 10
     # Wait for all pods to be in running state (no Terminating pods)
     while kubectl get pods -n red --no-headers | awk '$3' | grep -v "Running" > /dev/null; do sleep 1; done
@@ -31,6 +34,7 @@ function delete_create_trench () {
 
 function delete_create_trench_revert () {
     kubectl apply -f $(dirname -- $(readlink -fn -- "$0"))/configuration/init-trench-a.yaml
+    if [ ! $? -eq 0 ]; then exit 1 ; fi
     sleep 10
     kubectl wait --for=condition=Ready pods --all -n red --timeout=4m
 }
@@ -41,6 +45,7 @@ function new_stream () {
 
 function new_stream_revert () {
     kubectl delete stream -n red stream-a-iii
+    if [ ! $? -eq 0 ]; then exit 1 ; fi
     kubectl delete flow -n red flow-a-x-tcp
 }
 
@@ -50,6 +55,7 @@ function stream_max_targets () {
 
 function stream_max_targets_revert () {
     kubectl delete stream -n red stream-a-iii
+    if [ ! $? -eq 0 ]; then exit 1 ; fi
     kubectl delete flow -n red flow-a-x-tcp
 }
 
@@ -95,12 +101,14 @@ function flow_byte_matches_revert () {
 
 function new_attractor_nsm_vlan () {
     kubectl apply -f $(dirname -- $(readlink -fn -- "$0"))/configuration/new-attractor-nsm-vlan.yaml
+    if [ ! $? -eq 0 ]; then exit 1 ; fi
     sleep 5
     kubectl wait --for=condition=Ready pods --all -n red --timeout=4m
 }
 
 function new_attractor_nsm_vlan_revert () {
     kubectl delete -f $(dirname -- $(readlink -fn -- "$0"))/configuration/new-attractor-nsm-vlan.yaml
+    if [ ! $? -eq 0 ]; then exit 1 ; fi
     sleep 5
     while kubectl get pods -n red --no-headers | awk '$3' | grep -v "Running" > /dev/null; do sleep 1; done
 }
@@ -113,6 +121,44 @@ function conduit_destination_port_nats () {
 function conduit_destination_port_nats_revert () {
     kubectl apply -f $(dirname -- $(readlink -fn -- "$0"))/configuration/init-trench-a.yaml
     sleep 5
+}
+
+function kill_ipam () {
+    litmus "kill-ipam"
+}
+
+function kill_nsp () {
+    litmus "kill-nsp"
+}
+
+function kill_operator () {
+    litmus "kill-operator"
+}
+
+function kill_proxy () {
+    litmus "kill-proxy"
+}
+
+function kill_stateless_lb () {
+    litmus "kill-stateless-lb"
+}
+
+function kill_frontend () {
+    litmus "kill-frontend"
+}
+
+function litmus () {
+    kubectl apply -f $(dirname -- $(readlink -fn -- "$0"))/configuration/resiliency-$1.yaml
+    if [ ! $? -eq 0 ]; then exit 1 ; fi
+    sleep 5
+    # Wait for the restart
+    while [ "$(kubectl get ChaosEngine container-$1 -n litmus -o json | jq -r '.status.engineStatus')" != "completed" ] ; do sleep 1; done
+    # Checks the verdict
+    if [[ "$(kubectl get ChaosEngine container-$1 -n litmus -o json | jq -r '.status.experiments[] | select(.verdict != "Pass") | [.verdict] | @tsv' 2>&1 | wc -l)" != "0" ]]; then
+        >&2 echo "$1 failed"
+        exit 1
+    fi
+    kubectl delete -f $(dirname -- $(readlink -fn -- "$0"))/configuration/resiliency-$1.yaml
 }
 
 # Required to call the corresponding function
